ref: add regular ref content check for files backend
We implicitly rely on "git-fsck(1)" to check the consistency of regular
refs. However, when parsing the regular refs for files backend by using
"files-backend.c::parse_loose_ref_contents", we allow the ref content to
end with no newline or to contain some garbages.

Even though we never create such loose refs ourselves, we have accepted
such loose refs. So, it is entirely possible that some third-party tools
may rely on such loose refs being valid. We should not report an error
fsck message at current. But let's notice such a "curiously formatted"
loose refs being valid and tell the user our findings, so we can access
the possible extent of damage when we tighten the parsing rules in the
future.

And it's not suitable to either report a warn fsck message to the user.
This is because if the caller set the "strict" field in "fsck_options",
fsck warns will be automatically upgraded to errors. We should not allow
user to specify the "--strict" flag to upgrade the fsck warnings to
errors at current. It might cause compatibility issue which may break
the legacy repository. So we add the following two fsck infos to
represent the situation where the ref content ends without newline or has
garbages:

1. "refMissingNewline(INFO)": A ref does not end with newline. This kind
   of ref may be considered ERROR in the future.
2. "trailingRefContent(INFO)": A ref has trailing contents. This kind of
   ref may be considered ERROR in the future.

It may seem that we could not give the user any warnings by creating
fsck infos. However, in "fsck.c::fsck_vreport", we will convert
"FSCK_INFO" to "FSCK_WARN" and we can still warn the user about these
situations when using "git-refs verify" without introducing
compatibility issue.

In current "git-fsck(1)", it will report an error when the ref content
is bad, so we should following this to report an error to the user when
"parse_loose_ref_contents" fails. And we add a new fsck error message
called "badRefContent(ERROR)" to represent that a ref has a bad content.

In order to tell whether the ref has trailing content, add a new
parameter "trailing" to "parse_loose_ref_contents". Then introduce a new
function "files_fsck_refs_content" to check the regular refs to enhance
the "git-refs verify".

Mentored-by: Patrick Steinhardt <ps@pks.im>
Mentored-by: Karthik Nayak <karthik.188@gmail.com>
Signed-off-by: shejialuo <shejialuo@gmail.com>